using TabbedActivity everywhere with Generics; removed Routines Activity & Adapter, Practice&Learn Fragments, Video Adapter, etc
