+++ MyFitnessRoutines/src/main/java/com/devindow/myfitnessroutines/generic/GenericAdapter.java
@@ -1,4 +1,4 @@
-package com.devindow.myfitnessroutines.routine;
+package com.devindow.myfitnessroutines.generic;
 
 import android.content.Context;
 import android.graphics.Color;
@@ -12,15 +12,12 @@ import android.widget.TextView;
 
 import com.devindow.myfitnessroutines.BuildConfig;
 import com.devindow.myfitnessroutines.R;
+import com.devindow.myfitnessroutines.generic.Generic;
 import com.devindow.myfitnessroutines.routine.Routine;
 
 import java.util.ArrayList;
 
-/**
- * Created by Devin on 2/21/2018.
- */
-
-public class RoutineAdapter extends ArrayAdapter<Routine> {
+public class GenericAdapter extends ArrayAdapter<Generic> {
 
 	// Private Fields
 	private Context context;
@@ -28,8 +25,8 @@ public class RoutineAdapter extends ArrayAdapter<Routine> {
 
 
 	// Constructor
-	public RoutineAdapter(@NonNull Context context, int resource, @NonNull ArrayList<Routine> routines) {
-		super(context, resource, routines);
+	public GenericAdapter(@NonNull Context context, int resource, @NonNull ArrayList<Generic> generics) {
+		super(context, resource, generics);
 		this.context = context;
 		this.resource = resource;
 	}
@@ -40,32 +37,35 @@ public class RoutineAdapter extends ArrayAdapter<Routine> {
 	@Override
 	public View getView(int position, @Nullable View convertView, @NonNull ViewGroup parent) {
 
-		Routine routine = getItem(position);
+		Generic generic = getItem(position);
 
 		LayoutInflater inflater = LayoutInflater.from(context);
 		convertView = inflater.inflate(resource, parent, false);
 
 		// GRAY background for paid Routines in free app flavor
-		if (BuildConfig.FLAVOR.equals("free") && !routine.isFree) {
+		if (BuildConfig.FLAVOR.equals("free") && !generic.isFree) {
 			convertView.setBackgroundColor(Color.LTGRAY);
 		}
 
 		// GREEN background for Routines completed today
-		if (routine.ranRecently) {
+		if (generic.ranRecently) {
 			convertView.setBackgroundColor(Color.GREEN);
 		}
 
 		TextView txtName = convertView.findViewById(R.id.txtName);
-		txtName.setText(routine.name);
+		txtName.setText(generic.name);
 
-		TextView txtDuration = convertView.findViewById(R.id.txtDuration);
-		txtDuration.setText(routine.getTotalMinutesString());
+		if (generic instanceof Routine) {
+			Routine routine = (Routine)generic;
+			TextView txtDuration = convertView.findViewById(R.id.txtDuration);
+			txtDuration.setText(routine.getTotalMinutesString());
+		}
 
 		TextView txtDescription = convertView.findViewById(R.id.txtDescription);
-		if (routine.description.length() == 0) {
+		if (generic.description.length() == 0) {
 			txtDescription.setVisibility(View.GONE);
 		} else {
-			txtDescription.setText(routine.description);
+			txtDescription.setText(generic.description);
 		}
 
 		return convertView;
