+++ MyFitnessRoutines/src/main/java/com/devindow/myfitnessroutines/main/TabbedActivity.java
@@ -9,7 +9,6 @@ import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 
 import com.devindow.myfitnessroutines.R;
-import com.devindow.myfitnessroutines.main.SectionsPagerAdapter;
 
 public class TabbedActivity extends AppCompatActivity {
 
@@ -18,10 +17,10 @@ public class TabbedActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
 
         setContentView(R.layout.activity_tabbed);
-        
-        SectionsPagerAdapter sectionsPagerAdapter = new SectionsPagerAdapter(this, getSupportFragmentManager());
+
+        TabPagerAdapter tabPagerAdapter = new TabPagerAdapter(this, getSupportFragmentManager());
         ViewPager viewPager = findViewById(R.id.view_pager);
-        viewPager.setAdapter(sectionsPagerAdapter);
+        viewPager.setAdapter(tabPagerAdapter);
         TabLayout tabs = findViewById(R.id.tabs);
         tabs.setupWithViewPager(viewPager);
