+++ MyFitnessRoutines/src/main/res/layout/fragment_generic.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.devindow.myfitnessroutines.routine.RoutinesActivity"
+    tools:context="com.devindow.myfitnessroutines.main.TabbedActivity"
     tools:showIn="@layout/activity_main">
 
     <ListView
